fix: move transactions.ts file to right place
